+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/BuildDockerTask.java
@@ -24,6 +24,7 @@ import com.google.cloud.tools.jib.frontend.BuildStepsExecutionException;
 import com.google.cloud.tools.jib.frontend.BuildStepsRunner;
 import com.google.cloud.tools.jib.frontend.ExposedPortsParser;
 import com.google.cloud.tools.jib.frontend.HelpfulSuggestions;
+import com.google.cloud.tools.jib.frontend.ParameterValidator;
 import com.google.cloud.tools.jib.http.Authorization;
 import com.google.cloud.tools.jib.image.ImageReference;
 import com.google.cloud.tools.jib.image.InvalidImageReferenceException;
@@ -31,6 +32,7 @@ import com.google.cloud.tools.jib.registry.RegistryClient;
 import com.google.cloud.tools.jib.registry.credentials.RegistryCredentials;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
+import com.google.common.collect.ImmutableList;
 import javax.annotation.Nullable;
 import org.gradle.api.DefaultTask;
 import org.gradle.api.GradleException;
@@ -82,6 +84,19 @@ public class BuildDockerTask extends DefaultTask {
     GradleBuildLogger gradleBuildLogger = new GradleBuildLogger(getLogger());
     jibExtension.handleDeprecatedParameters(gradleBuildLogger);
 
+    ParameterValidator.checkListForNullOrEmpty(
+        jibExtension.getArgs(), "jib.container.args", gradleBuildLogger, GradleException::new);
+    ParameterValidator.checkListForNullOrEmpty(
+        jibExtension.getJvmFlags(),
+        "jib.container.jvmFlags",
+        gradleBuildLogger,
+        GradleException::new);
+    ParameterValidator.checkListForNullOrEmpty(
+        jibExtension.getExposedPorts(),
+        "jib.container.ports",
+        gradleBuildLogger,
+        GradleException::new);
+
     RegistryCredentials knownBaseRegistryCredentials = null;
     Authorization fromAuthorization = jibExtension.getFrom().getImageAuthorization();
     if (fromAuthorization != null) {
@@ -107,8 +122,8 @@ public class BuildDockerTask extends DefaultTask {
             .setBaseImageCredentialHelperName(jibExtension.getFrom().getCredHelper())
             .setKnownBaseRegistryCredentials(knownBaseRegistryCredentials)
             .setMainClass(mainClass)
-            .setJavaArguments(jibExtension.getArgs())
-            .setJvmFlags(jibExtension.getJvmFlags())
+            .setJavaArguments(ImmutableList.copyOf(jibExtension.getArgs()))
+            .setJvmFlags(ImmutableList.copyOf(jibExtension.getJvmFlags()))
             .setExposedPorts(
                 ExposedPortsParser.parse(jibExtension.getExposedPorts(), gradleBuildLogger))
             .setAllowHttp(jibExtension.getAllowInsecureRegistries());