+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/DockerContextMojo.java
@@ -18,8 +18,10 @@ package com.google.cloud.tools.jib.maven;
 
 import com.google.cloud.tools.jib.docker.DockerContextGenerator;
 import com.google.cloud.tools.jib.frontend.ExposedPortsParser;
+import com.google.cloud.tools.jib.frontend.ParameterValidator;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
+import com.google.common.collect.ImmutableList;
 import com.google.common.io.InsecureRecursiveDeleteException;
 import java.io.IOException;
 import java.nio.file.Paths;
@@ -52,6 +54,13 @@ public class DockerContextMojo extends JibPluginConfiguration {
     MavenBuildLogger mavenBuildLogger = new MavenBuildLogger(getLog());
     handleDeprecatedParameters(mavenBuildLogger);
 
+    ParameterValidator.checkListForNullOrEmpty(
+        getArgs(), "<container><args>", mavenBuildLogger, MojoExecutionException::new);
+    ParameterValidator.checkListForNullOrEmpty(
+        getJvmFlags(), "<container><jvmFlags>", mavenBuildLogger, MojoExecutionException::new);
+    ParameterValidator.checkListForNullOrEmpty(
+        getExposedPorts(), "<container><ports>", mavenBuildLogger, MojoExecutionException::new);
+
     Preconditions.checkNotNull(targetDir);
 
     MavenProjectProperties mavenProjectProperties =
@@ -65,10 +74,10 @@ public class DockerContextMojo extends JibPluginConfiguration {
 
       new DockerContextGenerator(mavenProjectProperties.getSourceFilesConfiguration())
           .setBaseImage(getBaseImage())
-          .setJvmFlags(getJvmFlags())
+          .setJvmFlags(ImmutableList.copyOf(getJvmFlags()))
           .setMainClass(mainClass)
-          .setJavaArguments(getArgs())
-          .setExposedPorts(getExposedPorts())
+          .setJavaArguments(ImmutableList.copyOf(getArgs()))
+          .setExposedPorts(ImmutableList.copyOf(getExposedPorts()))
           .generate(Paths.get(targetDir));
 
       mavenBuildLogger.lifecycle("Created Docker context at " + targetDir);